requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.5
djangorestframework==3.14.0
filelock==3.9.0
platformdirs==2.6.2
pytz==2022.7.1
sqlparse==0.4.3
virtualenv==20.17.1
